% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeCheck.R
\name{primeCheck}
\alias{primeCheck}
\title{Prime number checker}
\usage{
primeCheck(number)
}
\arguments{
\item{number}{The argument is named so because it is indeed a number that we would like to check whether it is prime or not.}
}
\value{
TRUE is number is prime, and FALSE otherwise.
}
\description{
This function takes a number as argument. It returns TRUE if it is a prime number, and FALSE otherwise.
}
\examples{
primeCheck(1)
primeCheck(3)
primeCheck(20)
primeCheck(101)

}
